fix: A single file to common stuff; A single compiler to all source files; These modifications prevent compilation issues.
