+++ HIP/CG/Makefile
@@ -1,19 +1,18 @@
 SHELL=/bin/sh
+
+CLASS=S
 BENCHMARK=cg
 BENCHMARKU=CG
+PARAMETERS=setparams
 
 include ../config/make.def
 
-OBJS = cg.o ${COMMON}/c_print_results.o  \
-       ${COMMON}/c_${RAND}.o ${COMMON}/c_timers.o ${COMMON}/c_wtime.o
+COMPILE_CG=$(CC) $(BENCHMARK).cpp -DCLASS_$(CLASS) -o $(BENCHMARK).$(CLASS)
+SET_PARAMETERS=$(CC) $(SYSDIR)/$(PARAMETERS).cpp -o $(PARAMETERS) && ./$(PARAMETERS) $(BENCHMARK) $(CLASS)
+MOVE_BINARIES=mv $(BENCHMARK).$(CLASS) $(BINDIR) && mv $(PARAMETERS) $(SYSDIR)
 
-include ../sys/make.common
-
-${PROGRAM}: config ${OBJS}
-	${CLINK} ${CLINKFLAGS} -o ${PROGRAM} ${OBJS} ${C_LIB}
-
-cg.o:             cg.cpp npbparams.hpp
-	${CCOMPILE} cg.cpp
+$(BENCHMARK):
+	$(pwd) $(SET_PARAMETERS) && $(COMPILE_CG) && $(MOVE_BINARIES)
 
 clean:
 	- rm -f *.o *~
