+++ HIP/FT/Makefile
@@ -1,20 +1,19 @@
 SHELL=/bin/sh
+
+CLASS=S
 BENCHMARK=ft
 BENCHMARKU=FT
+PARAMETERS=setparams
 
 include ../config/make.def
 
-OBJS = ft.o ${COMMON}/c_${RAND}.o ${COMMON}/c_print_results.o \
-       ${COMMON}/c_timers.o ${COMMON}/c_wtime.o
+COMPILE_FT=$(CC) $(BENCHMARK).cpp -DCLASS_$(CLASS) -o $(BENCHMARK).$(CLASS)
+SET_PARAMETERS=$(CC) $(SYSDIR)/$(PARAMETERS).cpp -o $(PARAMETERS) && ./$(PARAMETERS) $(BENCHMARK) $(CLASS)
+MOVE_BINARIES=mv $(BENCHMARK).$(CLASS) $(BINDIR) && mv $(PARAMETERS) $(SYSDIR)
 
-include ../sys/make.common
-
-${PROGRAM}: config ${OBJS}
-	${CLINK} ${CLINKFLAGS} -o ${PROGRAM} ${OBJS} ${C_LIB}
-
-ft.o:             ft.cpp npbparams.hpp
-	${CCOMPILE} ft.cpp
+$(BENCHMARK):
+	$(pwd) $(SET_PARAMETERS) && $(COMPILE_FT) && $(MOVE_BINARIES)
 
 clean:
-	- rm -f *.o *~ mputil*
-	- rm -f ft npbparams.hpp core
+	- rm -f *.o *~ 
+	- rm -f npbparams.hpp core
