+++ HIP/FT/ft.cpp
@@ -59,7 +59,7 @@
 
 #include <omp.h>
 #include <hip/hip_runtime.h>
-#include "../common/npb-CPP.hpp"
+#include "../common/npb.hpp"
 #include "npbparams.hpp"
 
 /*
