--- HIP/common/device_properties.cpp
@@ -1,73 +1,0 @@
-/* 
- * ------------------------------------------------------------------------------
- *
- * MIT License
- *
- * Copyright (c) 2021 Parallel Applications Modelling Group - GMAP
- *      GMAP website: https://gmap.pucrs.br
- *
- * Pontifical Catholic University of Rio Grande do Sul (PUCRS)
- * Av. Ipiranga, 6681, Porto Alegre - Brazil, 90619-900
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- * ------------------------------------------------------------------------------
- *
- * The original NPB 3.4 version was written in Fortran and belongs to:
- *      http://www.nas.nasa.gov/Software/NPB/
- *
- * ------------------------------------------------------------------------------
- *
- * The serial C++ version is a translation of the original NPB 3.4
- * Serial C++ version: https://github.com/GMAP/NPB-CPP/tree/master/NPB-SER
- *
- * Authors of the C++ code:
- *      Dalvan Griebler <dalvangriebler@gmail.com>
- *      Gabriell Araujo <hexenoften@gmail.com>
- *      Júnior Löff <loffjh@gmail.com>
- *
- * ------------------------------------------------------------------------------
- */
-
-#include <hip/hip_runtime.h>
-#include <stdio.h>
-#include <stdlib.h>
-
-#define DEVICE_NAME_INFO 0
-#define WARP_SIZE_INFO 1
-#define THREADS_PER_BLOCK_INFO 2
-
-int main(int argc, char** argv){
-    int gpu_device_id = atoi(argv[1]);
-    int functionality = atoi(argv[2]);
-
-    hipDeviceProp_t gpu_device_properties;
-    hipSetDevice(gpu_device_id);	
-    hipGetDeviceProperties(&gpu_device_properties, gpu_device_id);
-    
-    if(functionality==DEVICE_NAME_INFO){
-        printf("%s\n", gpu_device_properties.name);
-    } else if(functionality==WARP_SIZE_INFO){
-        printf("%d\n", gpu_device_properties.warpSize);
-    }else if(functionality==THREADS_PER_BLOCK_INFO){
-        printf("%d\n", gpu_device_properties.maxThreadsPerBlock);
-    }
-
-    return 0;
-}
