+++ HIP/sys/print_header
@@ -1,6 +1,6 @@
 echo '   ==========================================='
 echo '   =      NAS PARALLEL BENCHMARKS 4.1        ='
-echo '   =      CUDA Versions                      ='
+echo '   =      HIP Version                        ='
 echo '   =      C++                                ='
 echo '   ==========================================='
 echo ''
