+++ HIP/sys/print_instructions
@@ -1,13 +1,16 @@
 echo ''
-echo '   To make a NAS benchmark type '
+echo '   To make a NAS benchmark, type: '
 echo ''
 echo '         make <benchmark-name> CLASS=<class>'
 echo ''
-echo '   where <benchmark-name> is "cg", "ep", "ft", "is", or "mg"'
+echo '   where:' 
+echo '         <benchmark-name> is "cg", "ep", "ft", "is", "mg",'
+echo '                             "bt", "lu", or "sp"           ' 
 echo '         <class>          is "S", "W", "A", "B" or "C"'
 echo ''
-echo '   To make a set of benchmarks, create the file config/suite.def'
-echo '   according to the instructions in config/suite.def.template and type'
+echo '   To make a set of benchmarks, edit the file config/suite.def'
+echo '   according to the instructions in the file config/suite.def,'
+echo '   and type this command in the main directory:'
 echo ''
 echo '         make suite'
 echo ''
@@ -15,4 +18,3 @@ echo ' ***************************************************************'
 echo ' * Remember to edit the file config/make.def for site specific *'
 echo ' * information as described in the README file                 *'
 echo ' ***************************************************************'
-
